refactor(crypto): introduce StreamCipherAlgorithm base for stream ciphers
Add an abstract StreamCipherAlgorithm : SymmetricAlgorithm base that
centralizes the key/nonce validation, GenerateKey/IV, CreateEncryptor/
Decryptor, disposal, and ThrowIfDisposed boilerplate that ChaCha20 and
XChaCha20 previously duplicated. Subclasses now declare only their
sizes and a CreateStreamCipher factory.

Generalize the engine contract so each IStreamCipher owns its own
keystream position: GenerateKeystreamBlock(counter, dst) becomes
NextKeystreamBlock(dst), and the 32-bit counter plus overflow guard
move from the (now sealed, internal) StreamCipherTransform into
ChaCha20StreamCipher. This makes the abstraction uniform across ciphers
with a seekable block counter (ChaCha20, Salsa20) and those with an
evolving internal state and no random access (Rabbit, HC-128), so the
upcoming Salsa20/XSalsa20, Rabbit, and HC-128 ciphers plug into the
same base. The redundant per-algorithm ChaCha20Transform is removed.

Behaviour is unchanged: all published-vector KATs still pass and the
full crypto BVT suite is green (15,715 passed).
